backend/requirements.txt
# HerLuna Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic[email]
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4
scipy==1.14.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
pandas==2.2.3
